placeble.py and game_piece.py
All I did for placeble.py was add docstrings. 
For game_piece.py I struggled trying to create the is_valid_placement check for if the position is surrounded. I wasn't able to make it so ignore my code there keep yours. But I did make the equals method working. with the given example she has in step 2 to test I found out that she actually means "__eq__" for the method and not literally "equals".

is_valid_placement's test case just results in an error and I narrowed the error to be coming from one specific line of code in my version of is_valid_placement but even with that narrowed down I couldn't find the issue so I am going to email the professor.